+++ Repository/Generic/IGenericRepository.cs
@@ -3,10 +3,9 @@
     public interface IGenericRepository<T> where T : class
     {
         Task<T> Create(T item);
-        Task<T> Update(int id, T item);
+        Task<T> Update(long id, T item);
         Task<IEnumerable<T>> GetAll();
-        Task<T?> FindById(int id);
-        Task Delete(int id);
-
+        Task<T?> FindById(long id);
+        Task Delete(long id);
     }
 }
